Stroke fill works

+++ lottie/src/main/java/com/airbnb/lottie/LottieValueCallback.java
@@ -11,6 +11,8 @@ import android.support.annotation.RestrictTo;
  */
 @RestrictTo(RestrictTo.Scope.LIBRARY)
 public interface LottieValueCallback<T> {
-  T getValue(int startFrame, int endFrame, T startValue, T endValue, float linearKeyframeProgress,
-      float interpolatedKeyframeProgress, float overallProgress);
+  T getValue(
+      float startFrame, float endFrame,
+      T startValue, T endValue,
+      float linearKeyframeProgress, float interpolatedKeyframeProgress, float overallProgress);
 }
